+++ cli/src/subcommands/serve.rs
@@ -2,7 +2,7 @@ use crate::opts::ServeArgs;
 use crate::{create_execution_context, install_tracing_subscriber};
 use std::process::ExitCode;
 use tracing::{Level, event};
-use viceroy_lib::{Error, ViceroyService};
+use viceroy_lib::{Error, ProfilingConfig, ViceroyService};
 
 pub(crate) async fn exec(serve_args: ServeArgs) -> ExitCode {
     install_tracing_subscriber(serve_args.shared().verbosity());
@@ -29,15 +29,11 @@ pub(crate) async fn exec(serve_args: ServeArgs) -> ExitCode {
 /// Create a new server, bind it to an address, and serve responses until an error occurs.
 async fn serve(serve_args: ServeArgs) -> Result<(), Error> {
     // Load the wasm module into an execution context
-    let ctx = create_execution_context(
-        serve_args.shared(),
-        true,
-        serve_args.shared().guest_profile_config(),
-    )
-    .await?;
+    let ctx = create_execution_context(serve_args.shared(), true).await?;
 
-    if let Some(guest_profile_config) = serve_args.shared().guest_profile_config() {
-        std::fs::create_dir_all(guest_profile_config.path)?;
+    // Create profile directory if guest profiling is enabled
+    if let ProfilingConfig::Guest { path, .. } = serve_args.shared().profiling_config() {
+        std::fs::create_dir_all(path)?;
     }
 
     let addr = serve_args.addr();