+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add support for sending `PendingRequestHandle` downstream ([#631](https://github.com/fastly/Viceroy/pull/631))
 - Update to the latest WITs and adapter. ([#637](https://github.com/fastly/Viceroy/pull/637))
 - Fix to correctly apply TLS when using `wss:` URL schema with WebSockets passthrough ([#636](https://github.com/fastly/Viceroy/pull/636))
+- Internal unification of native/guest profiling config. (#[597](https://github.com/fastly/Viceroy/pull/597))
 
 ## 0.18.0 (2026-05-21)